Window Kernel module is buildable.

+++ CcKernel/Types/CcStatus.h
@@ -173,7 +173,7 @@ public:
   inline CcStatus& setSystemError(uint32 uiError)
     { m_eError = static_cast<EStatus>(static_cast<uint32>(EStatus::SystemError) | uiError); return *this; }
 #ifdef WINDOWS
-  inline CcStatus& setSystemError(unsigned long uiError)
+  inline CcStatus& setSystemError(unsigned int uiError)
     { m_eError = static_cast<EStatus>(uiError > static_cast<uint32>(EStatus::SystemError) ? uiError: ~uiError + 1); return *this; }
 #endif
